syndoc/pyproject.toml
[project]
name = "syndoc"
version = "0.1.0"
description = "Synopsys Document Parser - A scalable and extensible document parsing library inspired by docling"
license = {text = "MIT"}
keywords = [
    "document",
    "parser",
    "synopsys",
    "pdf",
    "layout-detection",
    "document-understanding",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pypdfium2>=4.0.0,<5.0.0",
    "pillow>=10.0.0,<11.0.0",
    "numpy>=1.24.0,<2.0.0",
    "torch>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["syndoc*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
